+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+# v1.0.1
+    -log Aws AccountId, UserId and Arn on successful connection test
+    -added Aws Profile info to filter message
+    -added ResourceList item properties IsFav and IsHidden
+    
 # v1.0.0
     -initial release